+++ examples/instance_with_advanced_options/main.tf
@@ -1,5 +1,5 @@
 /**
- * Copyright 2018 Google LLC
+ * Copyright 2019 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,44 +18,34 @@ provider "google" {
 }
 
 locals {
-  instance_name = format("%s-%s", var.instance_name, substr(md5(module.gce-container.container.image), 0, 8))
+  instance_name = format("%s-%s", var.instance_name, substr(md5(module.gce-advanced-container.container.image), 0, 8))
 }
 
-module "gce-container" {
+module "gce-advanced-container" {
   source = "../../"
 
-  cos_image_name = var.cos_image_name
-
   container = {
-    image = "gcr.io/google-samples/hello-app:1.0"
-
+    image = "busybox"
+    command = [
+      "tail"
+    ]
+    args = [
+      "-f",
+      "/dev/null"
+    ]
+    securityContext = {
+      privileged : true
+    }
+    tty : true
     env = [
       {
-        name  = "TEST_VAR"
-        value = "Hello World!"
-      },
-    ]
-
-    volumeMounts = [
-      {
-        mountPath = "/cache"
-        name      = "tempfs-0"
-        readOnly  = false
-      },
+        name  = "EXAMPLE"
+        value = "VAR"
+      }
     ]
   }
 
-  volumes = [
-    {
-      name = "tempfs-0"
-
-      emptyDir = {
-        medium = "Memory"
-      }
-    },
-  ]
-
-  restart_policy = "Always"
+  restart_policy = "OnFailure"
 }
 
 resource "google_compute_instance" "vm" {
@@ -66,7 +56,7 @@ resource "google_compute_instance" "vm" {
 
   boot_disk {
     initialize_params {
-      image = module.gce-container.source_image
+      image = module.gce-advanced-container.source_image
     }
   }
 
@@ -79,11 +69,11 @@ resource "google_compute_instance" "vm" {
   tags = ["container-vm-example"]
 
   metadata = {
-    gce-container-declaration = module.gce-container.metadata_value
+    gce-container-declaration = module.gce-advanced-container.metadata_value
   }
 
   labels = {
-    container-vm = module.gce-container.vm_container_label
+    container-vm = module.gce-advanced-container.vm_container_label
   }
 
   service_account {