+++ examples/managed_instance_group/variables.tf
@@ -52,13 +52,11 @@ variable "image_port" {
 variable "region" {
   description = "The GCP region to deploy instances into"
   type        = string
-  default     = "us-east4"
 }
 
 variable "zone" {
   description = "The GCP zone to deploy instances into"
   type        = string
-  default     = "us-east4-b"
 }
 
 variable "network" {