+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Extending the adopted spec, each change should have a link to its corresponding 
 
 ## [Unreleased]
 
+### Added
+
+- Example for showing advanced container options [#55]
+
 ## [2.0.0] - 2019-12-13
 
 ### Added
@@ -88,6 +92,7 @@ Extending the adopted spec, each change should have a link to its corresponding 
 [0.2.0]: https://github.com/terraform-google-modules/terraform-google-container-vm/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/terraform-google-modules/terraform-google-container-vm/compare/v0.1.0...v0.1.1
 
+[#55]: https://github.com/terraform-google-modules/terraform-google-container-vm/pull/55
 [#57]: https://github.com/terraform-google-modules/terraform-google-container-vm/pull/57
 [#54]: https://github.com/terraform-google-modules/terraform-google-container-vm/pull/54
 [#48]: https://github.com/terraform-google-modules/terraform-google-container-vm/pull/48